+++ DTCMS.IDAL/IDAL_Arc_Soft.cs
@@ -75,7 +75,7 @@ namespace DTCMS.IDAL
 		/// </summary>
 		/// <param name="count">返回记录数</param>
 		/// <returns>对象泛型集合</returns>
-		List<Arc_Soft> GetList(out long count);
+		List<Arc_Soft> GetList(out int count);
 
 		/// <summary>
 		/// 分页获取泛型数据列表
@@ -84,6 +84,6 @@ namespace DTCMS.IDAL
 		/// <param name="pageIndex">当前页数</param>
 		/// <param name="count">返回记录总数</param>
 		/// <returns>分页对象泛型集合</returns>
-		List<Arc_Soft> GetPageList(int pageSize, int pageIndex, out long count);
+		List<Arc_Soft> GetPageList(int pageSize, int pageIndex, out int count);
 	}
 }
