+++ DTCMS.IDAL/IDAL_AttachMent.cs
@@ -91,7 +91,7 @@ namespace DTCMS.IDAL
 		/// </summary>
 		/// <param name="count">返回记录数</param>
 		/// <returns>对象泛型集合</returns>
-		List<AttachMent> GetList(out long count);
+		List<AttachMent> GetList(out int count);
 
 		/// <summary>
 		/// 分页获取泛型数据列表
@@ -100,7 +100,7 @@ namespace DTCMS.IDAL
 		/// <param name="pageIndex">当前页数</param>
 		/// <param name="count">返回记录总数</param>
 		/// <returns>分页对象泛型集合</returns>
-		List<AttachMent> GetPageList(int pageSize, int pageIndex, out long count);
+		List<AttachMent> GetPageList(int pageSize, int pageIndex, out int count);
 
         /// <summary>
         /// 字符串缓存实现的通用分页存储过程
