+++ DTCMS.IDAL/IDAL_Users.cs
@@ -84,7 +84,7 @@ namespace DTCMS.IDAL
 		/// </summary>
 		/// <param name="count">返回记录数</param>
 		/// <returns>对象泛型集合</returns>
-		List<Users> GetList(out long count);
+		List<Users> GetList(out int count);
 
 		/// <summary>
 		/// 分页获取泛型数据列表
@@ -93,6 +93,6 @@ namespace DTCMS.IDAL
 		/// <param name="pageIndex">当前页数</param>
 		/// <param name="count">返回记录总数</param>
 		/// <returns>分页对象泛型集合</returns>
-		List<Users> GetPageList(int pageSize, int pageIndex, out long count);
+		List<Users> GetPageList(int pageSize, int pageIndex, out int count);
 	}
 }
